refactor(phase-2q): extract OpenProjectTimeEntryService
Continues the openproject_client.py god-class decomposition (ADR-002).
The four time-entry helpers move into a focused
``OpenProjectTimeEntryService`` exposed as ``self.time_entries``.
``OpenProjectClient`` keeps a thin delegator for each method so
existing callers work unchanged.

Methods moved
-------------
- ``get_time_entry_activities`` (file-based JSON read of all
  ``TimeEntryActivity.active``)
- ``create_time_entry`` (extracts WP/user/activity ids from the
  OpenProject API ``_embedded`` payload, builds a Ruby script with
  provenance-CF assignment, and calls ``execute_query_to_json_file``;
  ``escape_ruby_single_quoted`` is lazy-imported)
- ``get_time_entries`` (optional ``work_package_id`` / ``user_id``
  filters; uses a unique container filename per call to avoid
  concurrent-call collisions)
- ``batch_create_time_entries`` (writes a JSON payload to the
  container, runs a Rails runner script that creates ``TimeEntry``
  records in bulk, and reads structured results back via
  ``transfer_file_from_container``; per-entry failures are captured
  in the result rather than raising)

Honest docstring
----------------
``batch_create_time_entries``'s service-side docstring explicitly
states that partial success (``failed > 0`` alongside ``created > 0``)
is the normal shape — only Rails-execution / result-file failures
raise ``QueryExecutionError``. The pre-extraction docstring's
"Raises: QueryExecutionError if the batch operation fails" wording
was misleading because the Ruby loop already swallows per-entry
errors.

Numbers
-------
- ``openproject_client.py``: 3,984 -> 3,594 LOC (-390)
- ``openproject_time_entry_service.py``: 0 -> 495 LOC (new)
- Cumulative across phases 2a-2q: ``openproject_client.py``
  7,342 -> 3,594 LOC (-3,748, -51.0%)

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean (108 files)
- ruff check / ruff format: clean
